Improve float math and color conversion logic
Refactored Float32 conversion to always use bit_cast, removed legacy memcpy fallback. Updated fmod to use std::fmodf for float precision. Fixed color interpolation step calculation with explicit float cast. Simplified and clarified hue calculation in linearRGBToHSV for better readability and correctness.

+++ sdk/geUtilities/include/geMath.h
@@ -162,7 +162,7 @@ namespace geEngineSDK {
 
     GE_NODISCARD static FORCEINLINE float
     fmod(float X, float Y) {
-      return std::fmod(X, Y);
+      return std::fmodf(X, Y);
     }
 
     GE_NODISCARD static FORCEINLINE float
